requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
python-jose==3.3.0
python-multipart==0.0.6
authlib==1.2.1
httpx==0.25.0
itsdangerous==2.1.2
python-dotenv==1.0.0
jinja2==3.1.2 